+++ spec/resource/base_spec.rb
@@ -154,7 +154,7 @@ describe Served::Resource::Base do
 
       context 'response is success' do
         let(:body) { { attr2: 2, attr3: 3 }.stringify_keys }
-        let(:response) { double(code: 202, body: body.to_json) }
+        let(:response) { double(code: 200, body: body.to_json) }
         before do
           allow_any_instance_of(Served::HTTPClient).to receive(:delete).and_return(response)
         end